+++ js/background/background.js
@@ -330,16 +330,16 @@ async function processMessage(request, sender) {
 
         case 'closeWindow':
             windowId = cleanParameter(request.windowId);
-            if(windowId) {
-                try {
-                    const window = await chrome.windows.get(windowId);
-                    if (window) {
-                        await chrome.windows.remove(windowId);
-                        return true;
-                    }
-                } catch (error) {
-                    console.error("Error closing window:", error);
-                }
+            if (!windowId) {
+                return false;
+            }
+            
+            try {
+                await chrome.windows.get(windowId);
+                await chrome.windows.remove(windowId);
+                return true;
+            } catch (error) {
+                console.error("Error closing window:", error);
                 return false;
             }